[Gold I] Title: 도로검문, Time: 84 ms, Memory: 2436 KB -BaekjoonHub

+++ 백준/Gold/2307. 도로검문/도로검문.cc
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-using ll = long long;
+using ll = int;
 const ll INF = (ll)1e18;
 
 struct Edge {
